Refactor - fixing bad code design and adding support for swapping views.
Updated the module import file path to use a global variable declared by
main.ps1 on-load that identifies the root path of the program.

Clarified several variable names, and cleaned up the object construction
in ManagerConsole.psm1 and DataTreeView.psm1.

Added support in DataTreeView.psm1 for being able to dynamically load and
unload the components of the view.  DataListView.psm1 still needs to be
cleaned up and support added for setting data.